fix44: address reviewer findings — dead-code purge, IntArray widening, AI naming + file split
Code review identified critical + cleanup items pre-merge:

CRITICAL — byte-truncation latent bug
  • CascadeResult.triangleSegments was a ByteArray that stored
    leaf/component ids via .toByte(). On models with > 255 components
    (Dragon Scale, large multi-volume Bambu) the ids wrap silently.
  • Widened to IntArray across SegmentationCascade, CascadeResult,
    AiPaintResultState (.triangleSegments, .alternateTriangleSegments)
    and the one consuming unit test. The field has no production
    runtime readers today, but the carry-along value is now safe.

HIGH — dead code from churn (~330 LOC removed)
  • Deleted AiPaintViewModel.applyAiTopologyGrouping (78 LOC) — orphan
    from fix39, never called after fix39.3 simplified to "topology is
    deterministic via Z-banding, AI is just renaming".
  • Deleted TopologyRecursion.kt + its dedicated test — only used by
    the removed applyAiTopologyGrouping + the reverted fix40.3 K-means
    subdivision path.
  • Deleted MeshSegmenter.segmentByThickness (and the
    MeshThicknessComputer file it consumed) + segmentByBounds +
    coverageFractions — none referenced from production code paths.
  • Deleted AiLabelClient.labelGroups / buildGroupPrompt /
    parseGroupJson / GroupingResult / componentDisplayColors /
    hsvToArgb — only consumed by the dead applyAiTopologyGrouping.
  • Deleted RegionRow + ColourSwapSheet + MoveComponentSheet from
    AiPaintResultScreen (~200 LOC) — the bottom-sheet state setters
    (swapSheetRegion / moveSheetComponent) were never invoked from
    any composable, so the sheets never opened. Carry-over from the
    pre-fix34 tree shape.

HIGH — AI naming on topology primary (fix40.5 regression)
  • runPipeline was applying AI naming to cascadeResult.tree (paint
    state) but then primaryTree was being set to alternateResult.tree
    (topology) when swapToTopology fired — AI names were silently
    discarded for painted 3MFs. Decide-swap-first refactor now picks
    primaryCascade BEFORE running applyAiNaming so the AI-named tree
    is the one the user actually sees.

REFACTOR — split AiPaintResultScreen.kt (1147 → 455 LOC)
  • New AiPaintViewer.kt — 3D viewer composable + ViewerToolbar +
    SlotPaletteRow (~315 LOC). The viewer surface and its overlays
    moved here; the screen now wires it as a black-box.
  • New AiPaintHelpers.kt — viewLabelFor, computeFitCameraState,
    findNodeById, findLeafContainingTriangle, recenterForBed,
    AiNamingFailureChip (~140 LOC). All internal so they're reachable
    from both AiPaintResultScreen and AiPaintViewer.
  • AiPaintResultScreen.kt is now the Scaffold + state + tree wiring
    only — no GL viewer composable, no helper math.

Unit tests pass. Build clean. Two unused-param warnings remain on
AiPaintResultScreen's public API (onMoveComponent / onUpdateRegionColour)
— left in place for now since they're top-level params; addressing
them is a NavGraph + caller change.

+++ app/src/main/java/com/u1/slicer/aipaint/AiRegion.kt
@@ -31,7 +31,7 @@ data class AiPaintResultState(
     val trianglePositions: FloatArray = FloatArray(0),
     /** Per-triangle SEGMENT id (matches the cascade-tree leaf id that originally claimed the
      *  triangle; mutated only when the user does a "Clear all custom selections" flatten). */
-    val triangleSegments: ByteArray = ByteArray(0),
+    val triangleSegments: IntArray = IntArray(0),
     /** Per-triangle SLOT (0..3). Mutated by paint/lasso/cascade-reassign. Written to the 3MF. */
     val triangleRegions: ByteArray = ByteArray(0),
 
@@ -54,7 +54,7 @@ data class AiPaintResultState(
      *  switch — the alternate is a fresh start, not a layered overlay. */
     val alternateTree: List<AiRegionNode>? = null,
     val alternateSource: SegmentationSource? = null,
-    val alternateTriangleSegments: ByteArray? = null,
+    val alternateTriangleSegments: IntArray? = null,
 ) {
     override fun equals(other: Any?): Boolean = this === other
     override fun hashCode(): Int = System.identityHashCode(this)
